requirements.txt
flake8==5.0.4
flake8-docstrings==1.6.0
pyTelegramBotAPI==4.14.1
pytest==7.1.3
pytest-timeout==2.1.0
requests==2.26.0
python-dotenv==1.0.1